Improve diagnostic tooling Ctrl+C behavior.
Changing the cmd line tooling implementation in the following commit:

https://github.com/dotnet/diagnostics/commit/6837b8f319b08a458142605bcac76aff4e5af00a#diff-d33d6d23aba37ebcfc05462b2fa260136a11df1b03f129f1912a856c36a8d7f3

broke Ctrl+C handling in at least dotnet-trace, since it disabled
the process termination handler leading to process termination without
being able to close the session, creating a corrupt file.

dotnet-trace and dotnet-dsrouter tooling needs better control of
its Ctrl+C shutdown, so this commit implements a custom process
termination handler and disable the default handler implemented
by the cmd line tools. This handler makes sure to signal a
cancelation token if one of the configured signals hits the process.

This commit also fix an issue with the integrated execution of dsrouter
from other tools. Since Ctrl+C gets passed to child processes, integrated
dsrouter could stop before dotnet-trace was able to close the session
causing corrupt trace file and exception being logged in dotnet-trace.

To mitigate this, dsrouter implements a way to block signals only
accepting shutdown sequence on redirected stdin and dotnet-trace
will launch dsrouter in this mode. This makes the shutdown sequence
deterministic making sure dsrouter won't stop before dotnet-trace
completed any outstanding trace sessions.

+++ src/Tools/Common/DsRouterProcessLauncher.cs
@@ -36,7 +36,7 @@ namespace Microsoft.Internal.Common.Utils
 
         private Process ChildProc => _childProc;
 
-        public int Start(string dsroutercommand, CancellationToken ct)
+        public int Start(string dsrouterCommand, CancellationToken ct)
         {
             string toolsRoot = System.IO.Path.GetDirectoryName(System.Environment.ProcessPath);
             string dotnetDsrouterTool = "dotnet-dsrouter";
@@ -46,10 +46,13 @@ namespace Microsoft.Internal.Common.Utils
                 dotnetDsrouterTool = Path.Combine(toolsRoot, dotnetDsrouterTool);
             }
 
+            // Block SIGINT and SIGQUIT in child process.
+            dsrouterCommand += " --block-signals SIGINT;SIGQUIT";
+
             _childProc = new Process();
 
             _childProc.StartInfo.FileName = dotnetDsrouterTool;
-            _childProc.StartInfo.Arguments = dsroutercommand;
+            _childProc.StartInfo.Arguments = dsrouterCommand;
             _childProc.StartInfo.UseShellExecute = false;
             _childProc.StartInfo.RedirectStandardOutput = true;
             _childProc.StartInfo.RedirectStandardError = true;
@@ -78,7 +81,14 @@ namespace Microsoft.Internal.Common.Utils
             {
                 try
                 {
-                    _childProc.Kill();
+                    _childProc.StandardInput.WriteLine("Q");
+                    _childProc.StandardInput.Flush();
+
+                    _childProc.WaitForExit(1000);
+                    if (!_childProc.HasExited)
+                    {
+                        _childProc.Kill();
+                    }
                 }
                 // if process exited while we were trying to kill it, it can throw IOE
                 catch (InvalidOperationException) { }
